Test_Database_4/requirements.txt
# For Django app itself
boto3==1.34.108
celery>=5.3
channels>=4.0
channels-redis>=4.0
daphne>=4.0
Django==3.2.25
django-cors-headers==4.3
django-storages==1.14.2
djangorestframework==3.14.0
django-environ==0.11.2
django-s3-storage==0.15.0
djongo==1.3.6
dnspython==2.4.2
drf-spectacular==0.27.2
gunicorn==21.2.0
joblib==1.2.0
numpy>=1.26.0
pandas>=2.1.0
openpyxl==3.1.5
pymongo==3.12.1
psycopg2-binary==2.9.9
pypdf2==3.0.1
pytesseract==0.3.13
pytest==8.2.0
pytest-django==4.8.0
python-decouple==3.8
python-dotenv==0.21.0
pytz==2023.3.post1
python-dateutil==2.8.2
PyYAML>=6.0.1
redis==4.3.3
regex==2022.7.9
requests==2.31.0
responses==0.13.3
#scikit-learn==1.3.0
#scipy==1.11.1
seaborn==0.12.2
selenium==4.14.0
Send2Trash==1.8.0
service-identity==18.1.0
setuptools==68.0.0
sqlparse==0.2.4
uvicorn[standard]









